requirements.txt
python-dotenv
Flask requests python-dotenv